feat: startMaster, stopMaster, startViewer, stopViewer 기능 연결
